refactor: moved functions of event_utils.py to event.py
